backend/requirements.txt
flasgger==0.9.7.1
Flask==2.2.3
Flask_Cors==5.0.0
numpy==1.23.5
Pillow==9.4.0
Pillow==10.4.0
torch==1.13.1+cu117
torch==2.4.1
torchvision==0.14.1+cu117
torchvision==0.19.1
